Struct bear_lib_terminal::terminal::config::Input

pub struct Input {
    // some fields omitted
}

The input configuration section repr.

None values will not override current ones.

See terminal::set().

Methods

impl Input

fn empty() -> Input

Construct an input configuration section override segment with all elements equal to None.

fn precise_mouse(self, precise_mouse: bool) -> Self

Whether to generate a mouse-move event when a mouse moves from one pixel to another as opposed to from one cell to another.

Default: false.

fn mouse_cursor(self, mouse_cursor: bool) -> Self

Whether to show the cursor.

Default: true.

fn cursor_symbol(self, cursor_symbol: char) -> Self

The cursor symbol to blink in the read string function.

Default: '_' a.k.a. 0x5F.

fn cursor_blink_rate(self, cursor_blink_rate: i32) -> Self

Amount of time in milliseconds to blink the cursor symbol for.

Default: 500.