Function bear_lib_terminal::terminal::open

pub fn open(title: &str, width: u32, height: u32)

Creates the terminal window of the specified size with the specified title, without showing it. To show the window use the refresh() function.

Equivalent to the terminal_open() C API function with a subsequent call to the terminal_set() C API function with some defaults (NOTE: this will change).