Enum bear_lib_terminal::terminal::Event

pub enum Event {
    Close,
    Resize {
        width: i32,
        height: i32,
    },
    MouseMove {
        x: i32,
        y: i32,
    },
    MouseScroll {
        delta: i32,
    },
    KeyPressed {
        key: KeyCode,
        ctrl: bool,
        shift: bool,
    },
    KeyReleased {
        key: KeyCode,
        ctrl: bool,
        shift: bool,
    },
    ShiftPressed,
    ShiftReleased,
    ControlPressed,
    ControlReleased,
}

A single input event.

Variants
Close

Terminal window closed.

Resize

Terminal window resized. Needs to have window.resizeable = true to occur.

Note, that, as of 40e6253, the terminal window is cleared when resized.

Fields

width: i32

Width the terminal was resized to.

height: i32

Heigth the terminal was resized to.

MouseMove

Mouse moved.

If precise-mouse is off, generated each time mouse moves from cell to cell, otherwise, when it moves from pixel to pixel.

Fields

x: i32

0-based cell index from the left to which the mouse cursor moved.

y: i32

0-based cell index from the top to which the mouse cursor moved.

MouseScroll

Mouse wheel moved.

Fields

delta: i32

Amount of steps the wheel rotated.

Positive when scrolled "down"/"backwards".

Negative when scrolled "up"/"forwards"/"away".

KeyPressed

A keyboard or mouse button pressed (might repeat, if set in OS).

Fields

key: KeyCode

The key pressed.

ctrl: bool

Whether the Control key is pressed.

shift: bool

Whether the Shift key is pressed.

KeyReleased

A keyboard or mouse button released.

Fields

key: KeyCode

The key released.

ctrl: bool

Whether the Control key is pressed.

shift: bool

Whether the Shift key is pressed.

ShiftPressed

The Shift key pressed (might repeat, if set in OS).

ShiftReleased

The Shift key released.

ControlPressed

The Shift key pressed (might repeat, if set in OS).

ControlReleased

The Control key released.